criqndo service pedido pagamento

+++ src/main/java/com/bruno/domain/Produto.java
@@ -31,7 +31,7 @@ public class Produto implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	private String nome;
 	
@@ -53,7 +53,7 @@ public class Produto implements Serializable {
 	}
 	
 	
-	public Produto(Integer id, String nome,Double valor) {
+	public Produto(Long id, String nome,Double valor) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -72,11 +72,11 @@ public class Produto implements Serializable {
 		return Collections.emptyList();
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
